requirements.txt
aiohttp
beautifulsoup4
bilibili-api-python
browsermob_proxy
Flask
fuzzywuzzy
httpx
icalendar
jieba
loguru
lxml
numpy
pandas
Pillow
psutil
pydantic
PyMySQL
pypinyin
qrcode
requests
selenium
tinydb
tinyrecord
tqdm
ujson
apscheduler
matplotlib
wordcloud
mysql-connector-python
pyjsparser